+++ sdk/formrecognizer/ai-form-recognizer/src/index.ts
@@ -16,7 +16,7 @@ export {
   AddressValue,
   CopyAuthorization,
   CurrencyValue,
-  DocTypeInfo,
+  DocumentTypeDetails,
   DocumentBuildMode,
   DocumentFieldSchema,
   DocumentFieldType,
@@ -26,21 +26,24 @@ export {
   DocumentStyle,
   DocumentTableCellKind,
   LengthUnit,
-  DocumentModelInfo,
+  DocumentModelDetails,
   DocumentModelSummary,
   OperationSummary,
   OperationKind,
   OperationStatus,
-  ResourceInfo,
-  CustomDocumentModelsInfo,
+  ResourceDetails,
+  CustomDocumentModelsDetails,
   AnalyzeResultOperationStatus,
   SelectionMarkState,
   ParagraphRole,
-  DocumentPageKind,
   GetOperationResponse,
-  OperationInfo,
+  OperationDetails,
   ErrorModel,
   InnerError,
+  OperationDetailsUnion,
+  DocumentModelBuildOperationDetails,
+  DocumentModelComposeOperationDetails,
+  DocumentModelCopyToOperationDetails,
 } from "./generated";
 export {
   AnalysisPoller,
@@ -53,6 +56,5 @@ export {
 export { DocumentModelPoller, DocumentModelOperationState } from "./lro/administration";
 export * from "./models";
 export * from "./options";
-export * from "./prebuilt";
-export { EnglishCapitalLetter, Acronymic } from "./util";
+export * from "./documentModel";
 export { Point2D } from "./transforms/polygon";