+++ sdk/formrecognizer/ai-form-recognizer/src/lro/administration.ts
@@ -4,13 +4,19 @@
 import { PollOperationState, PollerLike } from "@azure/core-lro";
 import { OperationOptions } from "@azure/core-client";
 import { FormRecognizerError } from "../error";
-import { GetOperationResponse, DocumentModelInfo, OperationStatus } from "../generated";
+import {
+  DocumentModelDetails,
+  OperationStatus,
+  DocumentModelBuildOperationDetails,
+  DocumentModelCopyToOperationDetails,
+  DocumentModelComposeOperationDetails,
+} from "../generated";
 import { PollerOptions } from "../options/PollerOptions";
 
 /**
  * The state of a model creation operation.
  */
-export interface DocumentModelOperationState extends PollOperationState<DocumentModelInfo> {
+export interface DocumentModelOperationState extends PollOperationState<DocumentModelDetails> {
   /**
    * The status of the operation. One of:
    *
@@ -53,12 +59,19 @@ export interface DocumentModelOperationState extends PollOperationState<Document
   tags?: Record<string, string>;
 }
 
+// The generated type for GetOperationResult is not ideal here. This assertion is just kicking the can down the road but
+// it's about the only thing we can do to actually access the common `result` property.
+export type DocumentModelBuildResponse =
+  | DocumentModelBuildOperationDetails
+  | DocumentModelCopyToOperationDetails
+  | DocumentModelComposeOperationDetails;
+
 /**
  * Convert an operation result into a training poller state.
  * @internal
  */
 export async function toTrainingPollOperationState(
-  response: GetOperationResponse
+  response: DocumentModelBuildResponse
 ): Promise<DocumentModelOperationState> {
   return {
     operationId: response.operationId,
@@ -67,20 +80,20 @@ export async function toTrainingPollOperationState(
     percentCompleted: response.percentCompleted ?? 0,
     lastUpdatedOn: response.lastUpdatedDateTime,
     createdOn: response.createdDateTime,
-    result: response.result as DocumentModelInfo | undefined,
     error: response.error && new FormRecognizerError(response.error),
     isCancelled: response.status === "canceled",
     isCompleted: response.status === "succeeded",
     isStarted: response.status !== "notStarted",
     tags: response.tags,
+    result: response.result,
   };
 }
 
 /**
  * A long-running operation (poller) that tracks the state of a model creation operation, eventually producing a
- * {@link ModelInfo}.
+ * {@link DocumentModelDetails}.
  */
-export type DocumentModelPoller = PollerLike<DocumentModelOperationState, DocumentModelInfo>;
+export type DocumentModelPoller = PollerLike<DocumentModelOperationState, DocumentModelDetails>;
 
 /**
  * Defines a training operation.